Add SharedPreferences Singleton Service
This commit introduces a new file, `shared_preferences_singleton.dart`, within the `lib/core/services` directory. The file defines an abstract class `SharedPreferencesSingleton` that encapsulates the initialization and usage of the `SharedPreferences` package, providing a singleton pattern for accessing shared preferences throughout the application. The class includes static methods for initializing the shared preferences instance, setting boolean values, and retrieving boolean values with defaulting to `false` if the key is not found. This enhancement aims to simplify and centralize the management of shared preferences, promoting cleaner and more maintainable code.